requirements.txt
yt-dlp>=2024.1.0
sounddevice>=0.4.6
numpy>=1.24.0
pillow>=10.0.0
pystray>=0.19.4
